+++ lib/src/emoji_picker_internal_utils.dart
@@ -3,129 +3,40 @@ import 'dart:io';
 import 'dart:math';
 
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
-import 'package:emoji_picker_flutter/src/emoji_skin_tones.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'emoji_lists.dart' as emoji_list;
 import 'recent_emoji.dart';
+
+/// Initial value for RecentEmoji
+const initVal = 1;
 
 /// Helper class that provides internal usage
 class EmojiPickerInternalUtils {
   // Establish communication with native
   static const _platform = MethodChannel('emoji_picker_flutter');
-  static const _emojiVersion = 'emoji_version';
-
-  /// Returns true when local emoji list is outdated
-  Future<bool> isEmojiUpdateAvailable() async {
-    final prefs = await SharedPreferences.getInstance();
-    var emojiVersion = prefs.getInt(_emojiVersion) ?? 0;
-    // != to support downgrading of emoji_picker versions
-    return emoji_list.version != emojiVersion;
-  }
-
-  /// Updates local emoji version with current version
-  Future updateEmojiVersion() async {
-    final prefs = await SharedPreferences.getInstance();
-    prefs.setInt(_emojiVersion, emoji_list.version);
-  }
-
-  /// Restore locally cached emoji
-  Future<Map<String, String>?> _restoreFilteredEmojis(String title) async {
-    final prefs = await SharedPreferences.getInstance();
-    var emojiJson = prefs.getString(title);
-    if (emojiJson == null) {
-      return null;
-    }
-    var emojis =
-        Map<String, String>.from(jsonDecode(emojiJson) as Map<String, dynamic>);
-    return emojis;
-  }
 
   // Get available emoji for given category title
-  Future<Map<String, String>> _getAvailableEmojis(Map<String, String> map,
-      {required String title}) async {
-    Map<String, String>? newMap;
+  Future<CategoryEmoji> _getAvailableEmojis(CategoryEmoji category) async {
+    var available = (await _platform.invokeListMethod<bool>(
+        'getSupportedEmojis', {
+      'source': category.emoji.map((e) => e.emoji).toList(growable: false)
+    }))!;
 
-    // Get Emojis cached locally if available
-    if (await isEmojiUpdateAvailable() == false) {
-      newMap = await _restoreFilteredEmojis(title);
-    }
-
-    if (newMap == null) {
-      // Check if emoji is available on this platform
-      newMap = await _getPlatformAvailableEmoji(map);
-      // Save available Emojis to local storage for faster loading next time
-      if (newMap != null && title != Category.CUSTOM.name) {
-        await _cacheFilteredEmojis(title, newMap);
-      }
-    }
-
-    return newMap ?? {};
-  }
-
-  /// Returns map of all the available category emojis
-  Future<Map<Category, Map<String, String>>> getAvailableCategoryEmoji(
-      {required Map<String, String> customEmojis}) async {
-    final allCategoryEmoji = Map.fromIterables([
-      if (customEmojis.isNotEmpty) Category.CUSTOM,
-      Category.SMILEYS,
-      Category.ANIMALS,
-      Category.FOODS,
-      Category.ACTIVITIES,
-      Category.TRAVEL,
-      Category.OBJECTS,
-      Category.SYMBOLS,
-      Category.FLAGS
-    ], [
-      if (customEmojis.isNotEmpty) customEmojis,
-      emoji_list.smileys,
-      emoji_list.animals,
-      emoji_list.foods,
-      emoji_list.activities,
-      emoji_list.travel,
-      emoji_list.objects,
-      emoji_list.symbols,
-      emoji_list.flags,
+    return category.copyWith(emoji: [
+      for (int i = 0; i < available.length; i++)
+        if (available[i]) category.emoji[i]
     ]);
-
-    final futures = allCategoryEmoji.entries
-        .map((e) => _getAvailableEmojis(e.value, title: e.key.name));
-
-    final allAvailableEmojis = await Future.wait(futures);
-
-    return Map.fromIterables(allCategoryEmoji.keys, allAvailableEmojis);
   }
 
-  /// Stores filtered emoji locally for faster access next time
-  Future<void> _cacheFilteredEmojis(
-      String title, Map<String, String> emojis) async {
-    var emojiJson = jsonEncode(emojis);
-    final prefs = await SharedPreferences.getInstance();
-    prefs.setString(title, emojiJson);
-  }
-
-  /// Check if emoji is available on current platform
-  Future<Map<String, String>?> _getPlatformAvailableEmoji(
-      Map<String, String> emoji) async {
-    if (Platform.isAndroid) {
-      Map<String, String>? filtered = {};
-      var delimiter = '|';
-      try {
-        var entries = emoji.values.join(delimiter);
-        var keys = emoji.keys.join(delimiter);
-        var result = (await _platform.invokeMethod<String>('checkAvailability',
-            {'emojiKeys': keys, 'emojiEntries': entries})) as String;
-        var resultKeys = result.split(delimiter);
-        for (var i = 0; i < resultKeys.length; i++) {
-          filtered[resultKeys[i]] = emoji[resultKeys[i]]!;
-        }
-      } on PlatformException catch (_) {
-        filtered = null;
-      }
-      return filtered;
-    } else {
-      return emoji;
+  /// Filters out emojis not supported on the platform
+  Future<List<CategoryEmoji>> filterUnsupported(
+      List<CategoryEmoji> data) async {
+    if (kIsWeb || !Platform.isAndroid) {
+      return data;
     }
+    final futures = [for (final cat in data) _getAvailableEmojis(cat)];
+    return await Future.wait(futures);
   }
 
   /// Returns list of recently used emoji from cache
@@ -139,8 +50,37 @@ class EmojiPickerInternalUtils {
     return json.map<RecentEmoji>(RecentEmoji.fromJson).toList();
   }
 
-  /// Add an emoji to recently used list or increase its counter
+  /// Add an emoji to recently used list
   Future<List<RecentEmoji>> addEmojiToRecentlyUsed(
+      {required Emoji emoji, Config config = const Config()}) async {
+    // Remove emoji's skin tone in Recent-Category
+    if (emoji.hasSkinTone) {
+      emoji = removeSkinTone(emoji);
+    }
+    var recentEmoji = await getRecentEmojis();
+    var recentEmojiIndex =
+        recentEmoji.indexWhere((element) => element.emoji.emoji == emoji.emoji);
+    if (recentEmojiIndex != -1) {
+      // Already exist in recent list
+      // Remove it
+      recentEmoji.removeAt(recentEmojiIndex);
+    }
+    // Add it first position
+    recentEmoji.insert(0, RecentEmoji(emoji, initVal));
+
+    // Limit entries to recentsLimit
+    recentEmoji =
+        recentEmoji.sublist(0, min(config.recentsLimit, recentEmoji.length));
+
+    // save locally
+    final prefs = await SharedPreferences.getInstance();
+    prefs.setString('recent', jsonEncode(recentEmoji));
+
+    return recentEmoji;
+  }
+
+  /// Add an emoji to popular used list or increase its counter
+  Future<List<RecentEmoji>> addEmojiToPopularUsed(
       {required Emoji emoji, Config config = const Config()}) async {
     // Remove emoji's skin tone in Recent-Category
     if (emoji.hasSkinTone) {
@@ -153,14 +93,21 @@ class EmojiPickerInternalUtils {
       // Already exist in recent list
       // Just update counter
       recentEmoji[recentEmojiIndex].counter++;
+    } else if (recentEmoji.length == config.recentsLimit &&
+        config.replaceEmojiOnLimitExceed) {
+      // Replace latest emoji with the fresh one
+      recentEmoji[recentEmoji.length - 1] = RecentEmoji(emoji, initVal);
     } else {
-      recentEmoji.add(RecentEmoji(emoji, 1));
+      recentEmoji.add(RecentEmoji(emoji, initVal));
     }
+
     // Sort by counter desc
     recentEmoji.sort((a, b) => b.counter - a.counter);
+
     // Limit entries to recentsLimit
     recentEmoji =
         recentEmoji.sublist(0, min(config.recentsLimit, recentEmoji.length));
+
     // save locally
     final prefs = await SharedPreferences.getInstance();
     prefs.setString('recent', jsonEncode(recentEmoji));
@@ -168,21 +115,10 @@ class EmojiPickerInternalUtils {
     return recentEmoji;
   }
 
-  /// Returns true when the emoji support multiple skin colors
-  bool hasSkinTone(Emoji emoji) {
-    return emoji_list.supportSkinToneList.contains(emoji.emoji);
-  }
-
-  /// Applies skin tone to given emoji
-  Emoji applySkinTone(Emoji emoji, String color) {
-    final codeUnits = emoji.emoji.codeUnits;
-    var result = List<int>.empty(growable: true)
-      ..addAll(codeUnits.sublist(0, min(codeUnits.length, 2)))
-      ..addAll(color.codeUnits);
-    if (codeUnits.length >= 2) {
-      result.addAll(codeUnits.sublist(2));
-    }
-    return emoji.copyWith(emoji: String.fromCharCodes(result));
+  /// Clears the list of recent emojis in local storage
+  Future<void> clearRecentEmojisInLocalStorage() async {
+    final prefs = await SharedPreferences.getInstance();
+    prefs.setString('recent', jsonEncode([]));
   }
 
   /// Remove skin tone from given emoji
@@ -193,32 +129,5 @@ class EmojiPickerInternalUtils {
         '',
       ),
     );
-  }
-
-  static bool _isEnumItem(enumItem) {
-    final split_enum = enumItem.toString().split('.');
-    return split_enum.length > 1 &&
-        split_enum[0] == enumItem.runtimeType.toString();
-  }
-
-  static T? fromString<T>(List<T> enumValues, String value,
-      {bool camelCase = false}) {
-    try {
-      return enumValues.singleWhere((enumItem) =>
-          EmojiPickerInternalUtils.convertToString(enumItem,
-                  camelCase: camelCase)
-              .toLowerCase() ==
-          value.toLowerCase());
-    } on StateError catch (_) {
-      return null;
-    }
-  }
-
-  static String convertToString(dynamic enumItem, {bool camelCase = false}) {
-    assert(enumItem != null);
-    assert(_isEnumItem(enumItem),
-        '$enumItem of type ${enumItem.runtimeType.toString()} is not an enum item');
-    final _tmp = enumItem.toString().split('.')[1];
-    return _tmp;
   }
 }